# swagger_ui/retrieve/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Definition of ArmRay backend API for data retrieval operations"
  version: "0.0.1"
  title: "ArmRay API"
host: "ec2-52-36-38-196.us-west-2.compute.amazonaws.com:1151"
basePath: "/armray/api/0.1/rest/retrieve"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /Post/{id}:
    get:
      tags:
      - "Post"
      description: "Retrieves `Post` objects by **id**.\n"
      operationId: "getPostById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` name"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Post`"
          schema:
            $ref: "#/definitions/Post"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
  /Post/{id}/author:
    get:
      tags:
      - "Post"
      - "User"
      description: "Retrieves `Post` author\n"
      operationId: "getPostAuthorById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The author"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
  /Post/{id}/keys:
    get:
      tags:
      - "Post"
      description: "Retrieves `Post` keywords\n"
      operationId: "getPostKeywordsById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The keywords"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
  /Post/{id}/keys/{keyword}:
    get:
      tags:
      - "Post"
      description: "Retrieves whether keyword is on past\n"
      operationId: "isKeywordOnPost"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` id"
        required: true
        type: "string"
      - name: "keyword"
        in: "path"
        description: "the keyword"
        required: true
        type: "string"
      responses:
        200:
          description: "Boolean of whether the `User`is taged"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
  /Post/{id}/private:
    get:
      tags:
      - "Post"
      description: "Retrieves whether `Post` is private\n"
      operationId: "isPostPrivateById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
  /Post/{id}/tags:
    get:
      tags:
      - "Post"
      description: "Retrieves `Post` tags\n"
      operationId: "getPostTagsById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` name"
        required: true
        type: "string"
      responses:
        200:
          description: "The ids of the tagged `User` ids"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
  /Post/{id}/tags/{user_id}:
    get:
      tags:
      - "Post"
      - "User"
      description: "Retrieves whether `User`is tagged on past\n"
      operationId: "getIsUserTagOnProject"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` id"
        required: true
        type: "string"
      - name: "user_id"
        in: "path"
        description: "user_id to check for"
        required: true
        type: "string"
      responses:
        200:
          description: "Boolean of whether the `User`is taged"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
  /Post/{id}/text:
    get:
      tags:
      - "Post"
      description: "Retrieves `Post` text\n"
      operationId: "getPostTextById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The text"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
  /Post/{id}/title:
    get:
      tags:
      - "Post"
      description: "Retrieves `Post` title\n"
      operationId: "getPostTitleById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Post` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The title"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Post"
...
  /Project/{id}/posts:
    get:
      tags:
      - "Project"
      - "Post"
      description: "Get posts of a `Project`.\n"
      operationId: "getAllProjectPosts"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `Project` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The `reviews` associated with the `Project`"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Post"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Project"
  /Project/{id}/posts/{post_id}:
    get:
      tags:
      - "Project"
      - "Post"
      description: "Check if `Post` is on a `Project`\n"
      operationId: "isPostOnProject"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `Project` to retrieve"
        required: true
        type: "string"
      - name: "post_id"
        in: "path"
        description: "The `id` of the post"
        required: true
        type: "string"
      responses:
        200:
          description: "The post"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Project"
...
definitions:
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      fields:
        type: "string"
      message:
        type: "string"
...
  Post:
    type: "object"
    properties:
      created_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Post`creation"
      is_post_private:
        type: "boolean"
        description: "is `Post`private?"
      post_author:
        description: "author of post"
        $ref: "#/definitions/User"
      post_author_id:
        type: "string"
        description: "id of author of post"
      post_id:
        type: "string"
        description: "id of post"
      post_keywords:
        type: "array"
        description: "keywords for post"
        items:
          type: "string"
      post_tags:
        type: "array"
        description: "tagged users"
        items:
          $ref: "#/definitions/User"
      post_tags_id:
        type: "array"
        description: "tagged `User`ids"
        items:
          type: "string"
      post_text:
        type: "string"
        description: "text of post"
      post_title:
        type: "string"
        description: "title of post"
      updated_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Post`update"
...
  User:
    type: "object"
    properties:
      city:
        type: "string"
        description: "City of `User`location"
      created_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `User`creation"
      first_name:
        type: "string"
        description: "User's first name."
      geolocation:
        type: "array"
        description: "Tuple representing latitude/longitude of user"
        items:
          type: "number"
          format: "double"
      is_collaborator:
        type: "boolean"
        description: "Is `User`a `Project`collaborator?"
      is_creator:
        type: "boolean"
        description: "Is `User`a `Project`creator?"
      is_user_private:
        type: "boolean"
        description: "is `User`profile public or private?"
      keywords:
        type: "array"
        description: "array of keywords associated with the user"
        items:
          type: "string"
      last_name:
        type: "string"
        description: "User's last name"
      profile_images:
        type: "array"
        description: "Array of file system paths to profile images"
        items:
          type: "string"
      project_ids:
        type: "array"
        description: "array of `User`projects"
        items:
          type: "string"
      recent_collaborators_id:
        type: "array"
        description: "array of user_ids of recent collaborators"
        items:
          type: "string"
      review_ids:
        type: "array"
        description: "array of `Review`ids"
        items:
          type: "string"
      state:
        type: "string"
        description: "State of `User`location"
      university_ids:
        type: "array"
        description: "array of universities the `User`is associated with"
        items:
          type: "string"
      updated_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `User`update"
      user_id:
        type: "string"
        description: "Unique identifier representing a specific user"
      user_skill_ids:
        type: "array"
        description: "array of `UserSkill`ids"
        items:
          type: "string"
...